long unsigned data type

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	unsigned int a = 1, b = 1, c, count = 1;
+	long unsigned int a = 1, b = 1, c, count = 1;
 
 	while (count < 50)
 	{
